# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: CloudCarbon API
  description: |
    Open-source multi-cloud GreenOps platform for unified cost and carbon optimisation.

    **Authentication**: All endpoints (except `/health`) require either:
    - A `Bearer` JWT access token in the `Authorization` header, or
    - A `cc_`-prefixed API key in the `X-API-Key` header.

    **Session status**: Session 5 (final) — agent infrastructure, five autonomous agents,
    notification delivery, Prometheus metrics, structured logging, rate limiting, and
    production Helm chart are complete.
  version: "0.5.0"
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.cloudcarbon.io
    description: Production (future)

# =============================================================================
# Security
# =============================================================================
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# =============================================================================
# Tags
# =============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/token

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key prefixed with `cc_`, obtained from the dashboard

  # ---------------------------------------------------------------------------
  # Shared Parameters
  # ---------------------------------------------------------------------------
...
  schemas:
...
    EnrichmentRunRequest:
      type: object
      required: [tenant_id]
      properties:
        tenant_id:
          type: string
          format: uuid
          description: Tenant UUID to enrich records for
        mode:
          type: string
          enum: [incremental, full]
          default: incremental
          description: |
            incremental: only enrich records without an existing enriched_record.
            full: re-enrich all records (use when methodology is updated).
        enrichment_version:
          type: string
          nullable: true
          maxLength: 20
          description: Version string to stamp on re-enriched records (full mode only)
        limit:
          type: integer
          nullable: true
          minimum: 1
          maximum: 100000
          description: Maximum records to process (incremental mode only; default 1000)
    EnrichmentRunResponse:
      type: object
      required: [job_id, tenant_id, mode, status, queued_at]
      properties:
        job_id:
          type: string
          description: Job ID for status polling
        tenant_id:
          type: string
          format: uuid
        mode:
          type: string
          enum: [incremental, full]
        status:
          type: string
          example: queued
        queued_at:
          type: string
          format: date-time
    EnrichmentStatusResponse:
      type: object
      required: [job_id, status]
      properties:
        job_id:
          type: string
        status:
          type: string
          enum: [queued, running, completed, failed]
        processed:
          type: integer
          default: 0
        total:
          type: integer
          default: 0
        pct:
          type: number
          format: double
          minimum: 0
          maximum: 100
        failed:
          type: integer
          default: 0
        errors:
          type: array
          items:
            type: string
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        duration_seconds:
          type: number
          format: double
          nullable: true
    EnrichmentSummaryResponse:
      type: object
      required:
        - tenant_id
        - total_records_enriched
        - total_focus_records
        - coverage_pct
        - total_co2e_kg
        - total_scope3_co2e_kg
        - scope3_pct_of_total
        - total_water_litres
      properties:
        tenant_id:
          type: string
          format: uuid
        total_records_enriched:
          type: integer
          description: Number of FOCUS records with enrichment data
        total_focus_records:
          type: integer
          description: Total FOCUS records for the tenant
        coverage_pct:
          type: number
          format: double
          description: Percentage of FOCUS records with enrichment (0–100)
        total_co2e_kg:
          type: number
          format: double
          description: Total carbon across all scopes (kg CO2e)
        total_scope3_co2e_kg:
          type: number
          format: double
          description: Total Scope 3 carbon (kg CO2e)
        scope3_pct_of_total:
          type: number
          format: double
          description: Scope 3 as percentage of total carbon
        total_water_litres:
          type: number
          format: double
          description: Total water consumption (litres)
        last_enriched_at:
          type: string
          format: date-time
          nullable: true
        enrichment_version:
          type: string
          nullable: true
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          oneOf:
            - type: string
            - type: array
              items:
                type: object

# =============================================================================
# Paths
# =============================================================================
paths:
...
  /enrichment/run:
    post:
      tags: [Enrichment]
      summary: Start enrichment job
      description: |
        Trigger carbon and water enrichment for a tenant's FOCUS records.
        Runs as a background task; poll `/enrichment/status/{job_id}` for progress.

        **Modes:**
        - `incremental` (default): only enrich records without an existing enriched_record.
        - `full`: re-enrich all records (use after methodology updates).

        **Required role:** `engineer` or above.
      operationId: runEnrichment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrichmentRunRequest'
            examples:
              incremental:
                summary: Incremental enrichment (default)
                value:
                  tenant_id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                  mode: incremental
                  limit: 1000
              full_re_enrich:
                summary: Full re-enrichment with new methodology version
                value:
                  tenant_id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                  mode: full
                  enrichment_version: "1.1.0"
      responses:
        "202":
          description: Enrichment job queued successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrichmentRunResponse'
              example:
                job_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                tenant_id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                mode: incremental
                status: queued
                queued_at: "2026-04-01T12:00:00Z"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Insufficient role (requires engineer or above)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /enrichment/status/{job_id}:
    get:
      tags: [Enrichment]
      summary: Get enrichment job status
      description: |
        Poll the progress of an enrichment job by job_id.
        Job state is stored in Redis with a 24-hour TTL.
      operationId: getEnrichmentStatus
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
          description: Job ID returned by POST /enrichment/run
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrichmentStatusResponse'
              examples:
                running:
                  summary: Job in progress
                  value:
                    job_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                    status: running
                    processed: 342
                    total: 1000
                    pct: 34.2
                    failed: 0
                    errors: []
                    started_at: "2026-04-01T12:00:05Z"
                    completed_at: null
                    duration_seconds: null
                completed:
                  summary: Job completed
                  value:
                    job_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                    status: completed
                    processed: 1000
                    total: 1000
                    pct: 100.0
                    failed: 0
                    errors: []
                    started_at: "2026-04-01T12:00:05Z"
                    completed_at: "2026-04-01T12:02:37Z"
                    duration_seconds: 152.3
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Job not found or expired (24h TTL)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "503":
          description: Redis unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /enrichment/summary:
    get:
      tags: [Enrichment]
      summary: Get enrichment summary
      description: |
        Return aggregated carbon and water statistics for the current tenant.
        Includes total CO2e, Scope 3 breakdown, water consumption, and coverage percentage.

        **Required role:** `viewer` or above.
      operationId: getEnrichmentSummary
      parameters:
        - name: tenant_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: Tenant UUID
      responses:
        "200":
          description: Enrichment summary statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrichmentSummaryResponse'
              example:
                tenant_id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                total_records_enriched: 48230
                total_focus_records: 50000
                coverage_pct: 96.46
                total_co2e_kg: 12847.32
                total_scope3_co2e_kg: 8921.44
                scope3_pct_of_total: 69.44
                total_water_litres: 23104.8
                last_enriched_at: "2026-04-01T11:58:22Z"
                enrichment_version: "1.0.0"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ---------------------------------------------------------------------------
  # Recommendations (Session 4)
  # ---------------------------------------------------------------------------
...